[WebDriver][BiDi] Move Script and BrowsingContext to their own agents https://bugs.webkit.org/show_bug.cgi?id=291978
Reviewed by BJ Burg.

Extracted BrowsingContext and Script modules into separate agents - like BidiBrowserAgent -
to keep domain-specific logic out of the WebDriverBidiProcessor.

* Source/WebKit/Sources.txt:
* Source/WebKit/UIProcess/Automation/BidiBrowsingContextAgent.cpp: Added.
(WebKit::BidiBrowsingContextAgent::BidiBrowsingContextAgent):
(WebKit::BidiBrowsingContextAgent::activate):
(WebKit::BidiBrowsingContextAgent::close):
(WebKit::browsingContextPresentationFromCreateType):
(WebKit::BidiBrowsingContextAgent::create):
(WebKit::BidiBrowsingContextAgent::getTree):
(WebKit::BidiBrowsingContextAgent::handleUserPrompt):
(WebKit::pageLoadStrategyFromReadinessState):
(WebKit::BidiBrowsingContextAgent::navigate):
(WebKit::BidiBrowsingContextAgent::reload):
* Source/WebKit/UIProcess/Automation/BidiBrowsingContextAgent.h: Copied from Source/WebKit/UIProcess/Automation/WebDriverBidiProcessor.h.
* Source/WebKit/UIProcess/Automation/BidiScriptAgent.cpp: Added.
(WebKit::BidiScriptAgent::BidiScriptAgent):
(WebKit::BidiScriptAgent::callFunction):
(WebKit::BidiScriptAgent::evaluate):
* Source/WebKit/UIProcess/Automation/BidiScriptAgent.h: Added.
* Source/WebKit/UIProcess/Automation/WebAutomationSession.cpp: The notifiers
are called directly from the session, which eliminates null check fo the
session that was required before.

(WebKit::WebAutomationSession::willShowJavaScriptDialog):
(WebKit::WebAutomationSession::dismissCurrentJavaScriptDialog):
(WebKit::WebAutomationSession::acceptCurrentJavaScriptDialog):
(WebKit::WebAutomationSession::logEntryAdded):
* Source/WebKit/UIProcess/Automation/WebDriverBidiProcessor.cpp:
(WebKit::WebDriverBidiProcessor::WebDriverBidiProcessor):
(WebKit::WebDriverBidiProcessor::activate): Deleted.
(WebKit::WebDriverBidiProcessor::close): Deleted.
(WebKit::browsingContextPresentationFromCreateType): Deleted.
(WebKit::WebDriverBidiProcessor::create): Deleted.
(WebKit::WebDriverBidiProcessor::getTree): Deleted.
(WebKit::WebDriverBidiProcessor::handleUserPrompt): Deleted.
(WebKit::pageLoadStrategyFromReadinessState): Deleted.
(WebKit::WebDriverBidiProcessor::navigate): Deleted.
(WebKit::WebDriverBidiProcessor::reload): Deleted.
(WebKit::WebDriverBidiProcessor::userPromptOpenedOnPage): Deleted.
(WebKit::WebDriverBidiProcessor::userPromptClosedOnPage): Deleted.
(WebKit::WebDriverBidiProcessor::logEntryAdded): Deleted.
(WebKit::WebDriverBidiProcessor::callFunction): Deleted.
(WebKit::WebDriverBidiProcessor::evaluate): Deleted.
* Source/WebKit/UIProcess/Automation/WebDriverBidiProcessor.h:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/294128@main
