+++ Source/WebKit/UIProcess/Automation/WebDriverBidiProcessor.cpp
@@ -28,18 +28,11 @@
 
 #if ENABLE(WEBDRIVER_BIDI)
 
-// For AutomationError codes.
-#include "AutomationProtocolObjects.h"
 #include "BidiBrowserAgent.h"
+#include "BidiBrowsingContextAgent.h"
+#include "BidiScriptAgent.h"
 #include "Logging.h"
-#include "PageLoadState.h"
 #include "WebAutomationSession.h"
-#include "WebAutomationSessionMacros.h"
-#include "WebDriverBidiBackendDispatchers.h"
-#include "WebDriverBidiFrontendDispatchers.h"
-#include "WebDriverBidiProtocolObjects.h"
-#include "WebPageProxy.h"
-#include "WebProcessPool.h"
 #include <JavaScriptCore/InspectorBackendDispatcher.h>
 #include <JavaScriptCore/InspectorFrontendRouter.h>
 #include <wtf/TZoneMallocInlines.h>
@@ -48,11 +41,6 @@
 namespace WebKit {
 
 using namespace Inspector;
-using BrowsingContext = Inspector::Protocol::BidiBrowsingContext::BrowsingContext;
-using ReadinessState = Inspector::Protocol::BidiBrowsingContext::ReadinessState;
-using PageLoadStrategy = Inspector::Protocol::Automation::PageLoadStrategy;
-using UserPromptType = Inspector::Protocol::BidiBrowsingContext::UserPromptType;
-using UserPromptHandlerType = Inspector::Protocol::BidiSession::UserPromptHandlerType;
 
 WTF_MAKE_TZONE_ALLOCATED_IMPL(WebDriverBidiProcessor);
 
@@ -60,9 +48,9 @@ WebDriverBidiProcessor::WebDriverBidiProcessor(WebAutomationSession& session)
     : m_session(session)
     , m_frontendRouter(FrontendRouter::create())
     , m_backendDispatcher(BackendDispatcher::create(m_frontendRouter.copyRef()))
-    , m_browsingContextDomainDispatcher(BidiBrowsingContextBackendDispatcher::create(m_backendDispatcher, this))
-    , m_scriptDomainDispatcher(BidiScriptBackendDispatcher::create(m_backendDispatcher, this))
     , m_browserAgent(makeUnique<BidiBrowserAgent>(session, m_backendDispatcher))
+    , m_browsingContextAgent(makeUnique<BidiBrowsingContextAgent>(session, m_backendDispatcher))
+    , m_scriptAgent(makeUnique<BidiScriptAgent>(session, m_backendDispatcher))
     , m_browsingContextDomainNotifier(makeUnique<BidiBrowsingContextFrontendDispatcher>(m_frontendRouter))
     , m_logDomainNotifier(makeUnique<BidiLogFrontendDispatcher>(m_frontendRouter))
 {
@@ -118,274 +106,6 @@ void WebDriverBidiProcessor::sendBidiMessage(const String& message)
 void WebDriverBidiProcessor::sendMessageToFrontend(const String& message)
 {
     sendBidiMessage(message);
-}
-
-
-// MARK: Inspector::BidiBrowsingContextDispatcherHandler methods.
-
-void WebDriverBidiProcessor::activate(const BrowsingContext& browsingContext, CommandCallback<void>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    RefPtr webPageProxy = session->webPageProxyForHandle(browsingContext);
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!webPageProxy, WindowNotFound);
-
-    // FIXME: detect non-top level browsing contexts, returning `invalid argument`.
-    session->switchToBrowsingContext(browsingContext, emptyString(), [callback = WTFMove(callback)](CommandResult<void>&& result) {
-        if (!result) {
-            callback(makeUnexpected(result.error()));
-            return;
-        }
-
-        callback({ });
-    });
-}
-
-void WebDriverBidiProcessor::close(const BrowsingContext& browsingContext, std::optional<bool>&& optionalPromptUnload, CommandCallback<void>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: implement `promptUnload` option.
-    // FIXME: raise `invalid argument` if `browsingContext` is not a top-level traversable.
-
-    session->closeBrowsingContext(browsingContext);
-
-    callback({ });
-}
-
-static constexpr Inspector::Protocol::Automation::BrowsingContextPresentation defaultBrowsingContextPresentation = Inspector::Protocol::Automation::BrowsingContextPresentation::Tab;
-
-static Inspector::Protocol::Automation::BrowsingContextPresentation browsingContextPresentationFromCreateType(Inspector::Protocol::BidiBrowsingContext::CreateType createType)
-{
-    switch (createType) {
-    case Inspector::Protocol::BidiBrowsingContext::CreateType::Tab:
-        return Inspector::Protocol::Automation::BrowsingContextPresentation::Tab;
-    case Inspector::Protocol::BidiBrowsingContext::CreateType::Window:
-        return Inspector::Protocol::Automation::BrowsingContextPresentation::Window;
-    }
-
-    ASSERT_NOT_REACHED();
-    return defaultBrowsingContextPresentation;
-}
-
-void WebDriverBidiProcessor::create(Inspector::Protocol::BidiBrowsingContext::CreateType createType, const BrowsingContext& optionalReferenceContext, std::optional<bool>&& optionalBackground, const String& optionalUserContext, CommandCallback<BrowsingContext>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: implement `referenceContext` option.
-    // FIXME: implement `background` option.
-    // FIXME: implement `userContext` option.
-
-    session->createBrowsingContext(browsingContextPresentationFromCreateType(createType), [callback = WTFMove(callback)](CommandResultOf<BrowsingContext, Inspector::Protocol::Automation::BrowsingContextPresentation>&& result) {
-        if (!result) {
-            callback(makeUnexpected(result.error()));
-            return;
-        }
-
-        auto [resultContext, resultPresentation] = WTFMove(result.value());
-        callback(WTFMove(resultContext));
-    });
-}
-
-void WebDriverBidiProcessor::getTree(const BrowsingContext& optionalRoot, std::optional<double>&& optionalMaxDepth, CommandCallback<Ref<JSON::ArrayOf<Protocol::BidiBrowsingContext::Info>>>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: implement `root` option.
-    // FIXME: implement `maxDepth` option.
-
-    auto infos = JSON::ArrayOf<Inspector::Protocol::BidiBrowsingContext::Info>::create();
-    for (Ref process : session->protectedProcessPool()->processes()) {
-        for (Ref page : process->pages()) {
-            if (!page->isControlledByAutomation())
-                continue;
-
-            // FIXME: implement `parent` field.
-            // FIXME: implement `children` field.
-            // FIXME: implement `originalOpener` field.
-            // FIXME: implement `clientWindow` field.
-            // FIXME: implement `userContext` field.
-            infos->addItem(Inspector::Protocol::BidiBrowsingContext::Info::create()
-                .setContext(session->handleForWebPageProxy(page))
-                .setUrl(page->currentURL())
-                .setClientWindow("placeholder_window"_s)
-                .setUserContext("placeholder_context"_s)
-                .release());
-        }
-    }
-
-    callback({ { WTFMove(infos) } });
-}
-
-void WebDriverBidiProcessor::handleUserPrompt(const BrowsingContext& browsingContext, std::optional<bool>&& optionalShouldAccept, const String&, CommandCallback<void>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: implement `userText` option.
-
-    if (optionalShouldAccept && *optionalShouldAccept) {
-        callback(session->acceptCurrentJavaScriptDialog(browsingContext));
-        return;
-    }
-
-    // FIXME: this should consider the session's user prompt handler. <https://webkit.org/b/291666>
-    callback(session->dismissCurrentJavaScriptDialog(browsingContext));
-}
-
-
-// https://www.w3.org/TR/webdriver/#dfn-session-page-load-timeout
-static constexpr Seconds defaultPageLoadTimeout = 300_s;
-static constexpr ReadinessState defaultReadinessState = ReadinessState::None;
-
-static PageLoadStrategy pageLoadStrategyFromReadinessState(ReadinessState state)
-{
-    switch (state) {
-    case ReadinessState::None:
-        return PageLoadStrategy::None;
-    case ReadinessState::Interactive:
-        return PageLoadStrategy::Eager;
-    case ReadinessState::Complete:
-        return PageLoadStrategy::Normal;
-    }
-
-    ASSERT_NOT_REACHED();
-    return PageLoadStrategy::Normal;
-}
-
-void WebDriverBidiProcessor::navigate(const BrowsingContext& browsingContext, const String& url, std::optional<ReadinessState>&& optionalReadinessState, CommandCallbackOf<String, Inspector::Protocol::BidiBrowsingContext::Navigation>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    auto pageLoadStrategy = pageLoadStrategyFromReadinessState(optionalReadinessState.value_or(defaultReadinessState));
-    session->navigateBrowsingContext(browsingContext, url, pageLoadStrategy, defaultPageLoadTimeout.milliseconds(), [url, callback = WTFMove(callback)](CommandResult<void>&& result) {
-        if (!result) {
-            callback(makeUnexpected(result.error()));
-            return;
-        }
-
-        // FIXME: keep track of navigation IDs that we hand out.
-        callback({ { url, "placeholder_navigation"_s } });
-    });
-}
-
-void WebDriverBidiProcessor::reload(const BrowsingContext& browsingContext, std::optional<bool>&& optionalIgnoreCache, std::optional<ReadinessState>&& optionalReadinessState, CommandCallbackOf<String, Inspector::Protocol::BidiBrowsingContext::Navigation>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: implement `ignoreCache` option.
-
-    auto pageLoadStrategy = pageLoadStrategyFromReadinessState(optionalReadinessState.value_or(defaultReadinessState));
-    session->reloadBrowsingContext(browsingContext, pageLoadStrategy, defaultPageLoadTimeout.milliseconds(), [session = WTFMove(session), browsingContext, callback = WTFMove(callback)](CommandResult<void>&& result) {
-        if (!result) {
-            callback(makeUnexpected(result.error()));
-            return;
-        }
-
-        ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-        RefPtr webPageProxy = session->webPageProxyForHandle(browsingContext);
-        ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!webPageProxy, WindowNotFound);
-
-        // FIXME: keep track of navigation IDs that we hand out.
-        callback({ { webPageProxy->currentURL(), "placeholder_navigation"_s } });
-    });
-}
-
-void WebDriverBidiProcessor::userPromptOpenedOnPage(WebPageProxy& page, const UserPromptType& promptType, const UserPromptHandlerType& handlerType, const String& message, std::optional<String>&& defaultValue)
-{
-    RefPtr session = m_session.get();
-    if (!session)
-        return;
-
-    m_browsingContextDomainNotifier->userPromptOpened(session->handleForWebPageProxy(page), promptType, handlerType, message, defaultValue.value_or(emptyString()));
-}
-
-void WebDriverBidiProcessor::userPromptClosedOnPage(WebPageProxy& page, const UserPromptType& promptType, bool accepted, std::optional<String>&& userText)
-{
-    RefPtr session = m_session.get();
-    if (!session)
-        return;
-
-    m_browsingContextDomainNotifier->userPromptClosed(session->handleForWebPageProxy(page), promptType, accepted, userText.value_or(emptyString()));
-}
-
-// MARK: Log domain.
-
-void WebDriverBidiProcessor::logEntryAdded(const String& level, const String& source, const String& message, double timestamp, const String& type, const String& method)
-{
-    m_logDomainNotifier->entryAdded(level, source, message, timestamp, type, method);
-}
-
-// MARK: Inspector::BidiScriptDispatcherHandler methods.
-
-void WebDriverBidiProcessor::callFunction(const String& functionDeclaration, bool awaitPromise, Ref<JSON::Object>&& target, RefPtr<JSON::Array>&& arguments, std::optional<Protocol::BidiScript::ResultOwnership>&&, RefPtr<JSON::Object>&& optionalSerializationOptions, RefPtr<JSON::Object>&& optionalThis, std::optional<bool>&& optionalUserActivation, CommandCallbackOf<Protocol::BidiScript::EvaluateResultType, String, RefPtr<Protocol::BidiScript::RemoteValue>, RefPtr<Protocol::BidiScript::ExceptionDetails>>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: handle non-BrowsingContext obtained from `Target`.
-    std::optional<BrowsingContext> browsingContext = target->getString("context"_s);
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!browsingContext, InvalidParameter);
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session->webPageProxyForHandle(*browsingContext), WindowNotFound);
-
-    // FIXME: handle `awaitPromise` option.
-    // FIXME: handle `resultOwnership` option.
-    // FIXME: handle `serializationOptions` option.
-    // FIXME: handle custom `this` option.
-    // FIXME: handle `userActivation` option.
-
-    Ref<JSON::Array> argumentsArray = arguments ? arguments.releaseNonNull() : JSON::Array::create();
-
-    session->evaluateJavaScriptFunction(*browsingContext, emptyString(), functionDeclaration, WTFMove(argumentsArray), false, optionalUserActivation.value_or(false), std::nullopt, [callback = WTFMove(callback)](Inspector::CommandResult<String>&& result) {
-        auto evaluateResultType = result.has_value() ? Inspector::Protocol::BidiScript::EvaluateResultType::Success : Inspector::Protocol::BidiScript::EvaluateResultType::Exception;
-        auto resultObject = Inspector::Protocol::BidiScript::RemoteValue::create()
-            .setType(Inspector::Protocol::BidiScript::RemoteValueType::Object)
-            .release();
-
-        // FIXME: handle serializing different RemoteValue types as JSON.
-        if (result)
-            resultObject->setValue(JSON::Value::create(WTFMove(result.value())));
-
-        // FIXME: keep track of realm IDs that we hand out.
-        callback({ { evaluateResultType, "placeholder_realm"_s, WTFMove(resultObject), nullptr } });
-    });
-}
-
-void WebDriverBidiProcessor::evaluate(const String& expression, bool awaitPromise, Ref<JSON::Object>&& target, std::optional<Protocol::BidiScript::ResultOwnership>&&, RefPtr<JSON::Object>&& optionalSerializationOptions, std::optional<bool>&& optionalUserActivation, CommandCallbackOf<Protocol::BidiScript::EvaluateResultType, String, RefPtr<Protocol::BidiScript::RemoteValue>, RefPtr<Protocol::BidiScript::ExceptionDetails>>&& callback)
-{
-    RefPtr session = m_session.get();
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session, InternalError);
-
-    // FIXME: handle non-BrowsingContext obtained from `Target`.
-    std::optional<BrowsingContext> browsingContext = target->getString("context"_s);
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!browsingContext, InvalidParameter);
-    ASYNC_FAIL_WITH_PREDEFINED_ERROR_IF(!session->webPageProxyForHandle(*browsingContext), WindowNotFound);
-
-    // FIXME: handle `awaitPromise` option.
-    // FIXME: handle `resultOwnership` option.
-    // FIXME: handle `serializationOptions` option.
-
-    String functionDeclaration = makeString("function() {\n return "_s, expression, "; \n}"_s);
-    session->evaluateJavaScriptFunction(*browsingContext, emptyString(), functionDeclaration, JSON::Array::create(), false, optionalUserActivation.value_or(false), std::nullopt, [callback = WTFMove(callback)](Inspector::CommandResult<String>&& result) {
-        auto evaluateResultType = result.has_value() ? Inspector::Protocol::BidiScript::EvaluateResultType::Success : Inspector::Protocol::BidiScript::EvaluateResultType::Exception;
-        auto resultObject = Inspector::Protocol::BidiScript::RemoteValue::create()
-            .setType(Inspector::Protocol::BidiScript::RemoteValueType::Object)
-            .release();
-
-        // FIXME: handle serializing different RemoteValue types as JSON here.
-        if (result)
-            resultObject->setValue(JSON::Value::create(WTFMove(result.value())));
-
-        // FIXME: keep track of realm IDs that we hand out.
-        callback({ { evaluateResultType, "placeholder_realm"_s, WTFMove(resultObject), nullptr } });
-    });
 }
 
 } // namespace WebKit
